Use splitVarsByCondition method to codegen the left variables
